requirements.txt
Flask==2.2.5
requests==2.31.0
Pillow==10.0.0
schedule==1.2.0
